Start converting slurmctld to use create_mmap_buf().
Note one minor functional change in _open_part_state_file() -
no error will the thrown in the state file is < 10 bytes. I do not
believe this to be a big deal - the only such valid file that would
not immediate throw an error is 4-bytes containing just the
SLURM_PROTOCOL_VERSION and nothing else.